Add primitive error handling for MySQL errors.
We ran into a MySQL error that is causing our `generate_insert` to fail.
This adds naive error handling to the `generate_*` methods:
  1. it logs the table name and row id where the error occurs
  2. it handles the error so checksumo can proceed.

Executor has been updated to take an optional `on_fail` code block that's
executed in case of terminal failure: the value is returned as execute's return value.